Moving to testable chocomufin
